+++ packages/vue-component-library/src/stories/DropdownSingleSelect.spec.js
@@ -1,0 +1,10 @@
+describe('DropdownSingleSelect', () => {
+  it('Default', () => {
+    cy.visit(
+      '/iframe.html?id=dropdown-single-select--default&args=&viewMode=story'
+    )
+    cy.get('.dropdown-single-select').should('exist')
+
+    cy.percySnapshot('DropdownSingleSelect: Default')
+  })
+})